Remove custom override for PREUPLOAD.cfg
The custom hooks configuration included certain hooks we don't need
The hook unit tests are not runnable on a standard python environment,
so removing them now from the hooks.

The hook changes will be tested immediately on uploading themselves,
so removing the unit tests from hooks doesn't seem to compromise a lot.

Test: Global hooks config for studio-master-dev should be used for this
repo too, which means it will test for this 'Test:' field in this and
subsequent CLs

Change-Id: I262d5ace2e80ddf0c19f258f21d9c4bb941469d7