Pass --skip-sorting-imports to google-java-format.
In the newest version of the formatter, imports are sorted by default.
This would normally be fine, except that AOSP has a defined order that
doesn't confirm to this style, which would lead to the formatter
requiring a lot of changes and possibly introducing conflicts without
a larger scale cleanup. So turn off this behavior.

Projects that would prefer to sort their imports may specify the
--sort-options arg in their PREUPLOAD.cfg files to revert to the
default google-java-format behavior.

Bug: 31552314
Test: Verified in local test repository
Change-Id: I9e2df57ef472e7f2d1d4e795ba63559c1aacac67