Add preupload hook to make sure IDE & PSQ profiles stay synced.
Every inspection enabled on PSQ should be enabled locally in the IDE
as well. This hook ensures this by tracking if any of the commits being
uploaded for review touches either IDE or PSQ profiles, and compares
them if so, ensuring that PSQ profile is a subset of the local IDE one,
with the same inspection enabled and having identical settings.

If any inconsistencies are found, the hook will report them, fix them
in the working tree, and fail. The auto-fixed profiles will need to be
re-added and re-committed in order for the upload to proceed.

Test: Day-to-day usage tool.
Bug: b/62258402
Change-Id: I6babfbcf5203f26ce06f9bd1f12389ce8089e56f