Add check for duplicates, and sort before dumping the auto-corrected profile.
IDEA default layout is such that there are no duplicate
inspection classes in the profile, and that the inspections
are sorted in alphabetical order.

This CL enhances the sync-inspection-profiles hook to always
check for duplicates and dump the auto-corrected profile in
the sorted form.

Test: Verified by regenerating Studio & Tools profiles and
seeing they ahve the intended contents.

Change-Id: I043f8401dd643287b03d7042f621bed6faeeb233