Patch ElementTree to preserve original relative ordering of attributes.
This is needed to make sure that whenever hook applies its autofixes to
the IDE profile, the diff doesn't explode with a lot of irrelevant line
changes caused by just random attributes ordering.

NB: This CL creates a local (patched) version of ElementTree/ElementPath
from Python's xml.etree.ElementTree. This one should always suit the
purposes of the hook, but if needed, comparisons with stdlib
can be done to see how the patch can be re-applied. It's unlikely
though, because those files haven't been changed in stdlib for 10 years
now.

Test: N/A
Change-Id: I1ba91d5eeb0840cb8663d667ff8166e1a563b71d