Rebase 'studio-master-dev' on 'master' and squash
studio-master-dev has some customizations on top of AOSP master so we
cannot simply track AOSP master. This change rebase studio-master-dev
on top of the tip of AOSP master while preserving customizations in
studio-master-dev and resolving conflicts along the way.

The change essentially updates all these scripts to be runnable with
python3, which is required with the latest repo tool.

Bug: 149266692
Test: manual
Change-Id: Ic21520839ab47a7b14b961b9799c89ca0f1509f3