Fix pre-upload hook on Windows (python 2.7)
Using str.decode converts the resulting string into unicode on Windows.
This is not expected, as most strings are expected to be non-unicode when
running scripts with Python 2.7. In this case, the strings returned
by "run_command" sometimes end up in the environment variables used to
create a sub-process, and Python 2.7 does not support unicode strings
for the process creation API.

Calling "str()" on the result of "decode()" ensures the string is converted
back to an 8-bit string. On non-Windows platform (and on Python 3+), the
call is a no-op.

This issue was introduced by a recent rebase/merge of the repohooks
repository.

Test: Manual (on a Windows machine)
Bug: 150001153
Bug: 149406150
Change-Id: I67f4c7ee9700c8082e3d965a3dc8ec5bec91955b