Send clang-format path when invoking aidl-format
aidl-format internally uses clang-format, but so far has been relying on
PATH to locate it. Therefore, the tool didn't work if the user didn't
install clang-format on the host machine which sometimes isn't possible
(due to the lack of privilege).

This change fixes the problem by passing the path to AOSP-hosted
clang-format binary via a new argument --clang-format-path.

Bug: 254457261
Test: rename /usr/bin/clang-format to something else
make an ugly change in any AIDL file, create a commit, do the repo
upload. The ugly change is detected and fixup is also suggested.

Change-Id: I87271d8b931092b33525dc5aea02ed5f83442629