pre-upload: include "hook" in the error summary
When showing errors to people, it can be confusing to just see the
plain hook name without any context.  For example, if bpfmt fails,
we see:
	[FAILED] bpfmt
	...

This might make people think the bpfmt tool itself failed, or if
they aren't familiar with what bp files are, this short output can
be a little vague.  Add the word "hook" to the end to make it a bit
more clear that a specific hook is failing.
	[FAILED] bpfmt hook
	...

Bug: None
Test: unittests
Change-Id: I86cfd603f612121a655f45e88b4e002071a85f0e