hooks: gofmt: fix argument handling
Since gofmt iterates over the filtered set of files by hand, and
reads it from the git commit to run through the formatter, we don't
want to expand the list of files into the base command.  If they
end up there, gofmt will run on them as they exist on disk, not on
the content as they exist in the git commit.

Also make sure to pass the set of custom gofmt arguments to the
fixup command so it gets run correctly.

Bug: 116112526
Test: unittests
Change-Id: I3615afa97817482c4e0482053fa308bed8c8d12b