pre-upload: flip hook progress output style
Currently the hook status line shows how many hooks have been run,
and what the current hook being executed is.  This is fine for a
single-threaded model, but is confusing if we want to run hooks in
parallel.  So flip the status to show all the running and pending
hooks.  This is a bit more verbose, but it's the only way to really
represent the work that is being done.

In practice, many of the hooks are very fast, so the long line will
collapse very quickly to something more manageable.  If the line is
too long for the current terminal, we will chop it.

Old:
[RUNNING 2/13] hooks_unittest
New:
[RUNNING 4/13] shell_unittest, config_unittest, utils_unittest, ...

Bug: None
Test: `./pre-upload.py` output
Change-Id: I280682c93ae4164adf210ba5d4a54867af377e85