pylint: support multiple pylintrc files per-repo
For a python file, run pylint with the pylintrc file that is found
in the same directory or in any parent directory in the repo.
If no matching pylintrc file can be found then fallback to default
pylintrc file.

Any pylint options specified in PREUPLOAD.cfg will be applied to
all pylint runs.

Bug: 342197112
Test: add pylint trigger to platform/build git, modify python files
    and confirm that pylint is executed with correct configuration
Change-Id: I1bda8a491cff5b1f5cfd79b40969210cc417688e